added Recipe Config to initialize recipe table with data

+++ src/main/java/com/prepper/prepper/controller/RecipesController.java
@@ -27,7 +27,7 @@ public class RecipesController {
 
     @DeleteMapping(path = "{recipeId}")
     public void removeRecipe(
-            @PathVariable("recipeId")Long recipeId) {
+            @PathVariable("recipeId")Integer recipeId) {
         recipeService.removeRecipe(recipeId);
     }
 }
